protect basic contact types

+++ type_contact/view.php
@@ -1,41 +1,30 @@
 <?php
 ob_start();
 $rootpath = "../";
+$role = 'admin';
 require_once($rootpath."includes/inc_default.php");
 require_once($rootpath."includes/inc_adoconnection.php");
-session_start();
-$s_id = $_SESSION["id"];
-$s_name = $_SESSION["name"];
-$s_letscode = $_SESSION["letscode"];
-$s_accountrole = $_SESSION["accountrole"];
+
+if(!isset($s_id) && ($s_accountrole == "admin")){
+	header('Location: ' . $rootpath . 'login.php');
+	exit;
+}
+
+if (!isset($_GET["id"])){
+	header("Location: overview.php");
+	exit;
+}
+
+$id = $_GET["id"];
+$contacttype = get_contacttype($id);
 
 include($rootpath."includes/inc_header.php");
-include($rootpath."includes/inc_nav.php");
+echo "<h1>Contacttype bekijken</h1>";
+show_contacttype($contacttype);
+include($rootpath."includes/inc_footer.php");
 
-if(isset($s_id) && ($s_accountrole == "admin")){
-	if (isset($_GET["id"])){
-		$id = $_GET["id"];
-		$contacttype = get_contacttype($id);
-		show_ptitle();
-		show_contacttype($contacttype);
-	}else{
-		redirect_overview();
-	}
-}else{
-	redirect_login($rootpath);
-}
 
-////////////////////////////////////////////////////////////////////////////
-//////////////////////////////F U N C T I E S //////////////////////////////
-////////////////////////////////////////////////////////////////////////////
-
-function redirect_login($rootpath){
-	header("Location: ".$rootpath."login.php");
-}
-
-function show_ptitle(){
-	echo "<h1>Contacttype bekijken</h1>";
-}
+///////////////////////
 
 function show_contacttype($contacttype){
 	echo "<div >";
@@ -54,10 +43,18 @@ function show_contacttype($contacttype){
 	echo "</td>";
 	echo "</tr>";
 	echo "</table></div>";
+	
+		echo "<div class='border_b'>";
+	if (in_array($contacttype['abbrev'], array('mail', 'tel', 'gsm', 'adr')))
+	{
+		echo '<p>Beschermd contact type: kan niet aangepast worden.</p>';
+	}
+	else
+	{
+		echo "| <a href='edit.php?id=".$contacttype["id"]."'>Aanpassen</a> | ";
+		echo "<a href='delete.php?id=".$contacttype["id"]."'>Verwijderen</a> |";
 
-	echo "<div class='border_b'>";
-	echo "| <a href='edit.php?id=".$contacttype["id"]."'>Aanpassen</a> | ";
-	echo "<a href='delete.php?id=".$contacttype["id"]."'>Verwijderen</a> |";
+	}
 	echo "</div>";
 }
 
@@ -68,11 +65,3 @@ function get_contacttype($id){
 	$contacttype = $db->GetRow($query);
 	return $contacttype;
 }
-
-function redirect_overview(){
-	header("Location: overview.php");
-}
-
-include($rootpath."includes/inc_sidebar.php");
-include($rootpath."includes/inc_footer.php");
-?>
